backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "worton-esg-backend"
version = "0.1.0"
description = "FastAPI backend skeleton for the Daton ESG platform."
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16,<2.0",
  "asyncpg>=0.30,<1.0",
  "boto3>=1.40,<2.0",
  "fastapi>=0.115,<1.0",
  "greenlet>=3.2,<4.0",
  "openai>=1.109.1,<2.0",
  "openpyxl>=3.1.5,<4.0",
  "pdfplumber>=0.11.7,<1.0",
  "pinecone>=7.3.0,<8.0",
  "PyJWT[crypto]>=2.10,<3.0",
  "python-docx>=1.2.0,<2.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic-settings>=2.4,<3.0",
  "sqlalchemy>=2.0.36,<3.0",
  "structlog>=24.4,<26.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.15.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
